requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.29.0
redis[hiredis]==5.0.4
httpx==0.27.0
openai==1.35.7
sentence-transformers==3.0.1
rank-bm25==0.2.2
nltk==3.8.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-fastapi-instrumentator==6.1.0
structlog==24.2.0
aiofiles==23.2.1
